+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in CUPS v2.4.8 (TBA)
 - Added new value for 'lpstat' option '-W' - successfull - for getting
   successfully printed jobs (Issue #830)
 - Added support for PAM modules password-auth and system-auth (Issue #892)
+- Updated IPP Everywhere printer creation error reporting (Issue #347)
 - Raised `cups_enum_dests()` timeout for listing available IPP printers
   (Issue #751)
 - Fixed the web interface not showing an error for a non-existent printer